+++ src/de/hhu/bsinfo/utils/reflect/dt/DataTypeParserInt.java
@@ -23,11 +23,11 @@ public class DataTypeParserInt implements DataTypeParser
 		{
 			String tmp = p_str.substring(0, 2);
 			if (tmp.equals("0x"))
-				return java.lang.Integer.parseInt(p_str.substring(2), 16);
+				return (int) java.lang.Long.parseLong(p_str.substring(2), 16);
 			else if (tmp.equals("0b"))
-				return java.lang.Integer.parseInt(p_str.substring(2), 2);
+				return (int) java.lang.Long.parseLong(p_str.substring(2), 2);
 			else if (tmp.equals("0o"))
-				return java.lang.Integer.parseInt(p_str.substring(2), 8);
+				return (int) java.lang.Long.parseLong(p_str.substring(2), 8);
 		}
 		
 		return java.lang.Integer.parseInt(p_str);
